Rename oidcescape to oidcmapping
