+++ examples/oidc-mapping/gateway.toml
@@ -11,7 +11,7 @@ commit_share_to_storage_ref = true
 home_provider = "/home"
 
 [grpc.services.authregistry.drivers.static.rules]
-oidcescape = "localhost:13000"
+oidcmapping = "localhost:13000"
 
 [grpc.services.storageregistry.drivers.static.rules."/home"]
 address = "localhost:17000"
