+++ examples/oidc-mapping/users-oidcmapping.toml
@@ -8,10 +8,10 @@ jwt_secret = "Pive-Fumkiu4"
 address = "0.0.0.0:13000"
 
 [grpc.services.authprovider]
-auth_manager = "oidcescape"
+auth_manager = "oidcmapping"
 [grpc.services.authprovider.auth_managers.json]
 users = "users.json"
-[grpc.services.authprovider.auth_managers.oidcescape]
+[grpc.services.authprovider.auth_managers.oidcmapping]
 issuer = "http://iam-login-service:8080/"
 userprovidersvc = "0.0.0.0:13000"
 # The ESCAPE IAM users mapping file path
