+++ pkg/auth/manager/loader/loader.go
@@ -28,7 +28,7 @@ import (
 	_ "github.com/cs3org/reva/pkg/auth/manager/machine"
 	_ "github.com/cs3org/reva/pkg/auth/manager/nextcloud"
 	_ "github.com/cs3org/reva/pkg/auth/manager/oidc"
-	_ "github.com/cs3org/reva/pkg/auth/manager/oidcescape"
+	_ "github.com/cs3org/reva/pkg/auth/manager/oidcmapping"
 	_ "github.com/cs3org/reva/pkg/auth/manager/owncloudsql"
 	_ "github.com/cs3org/reva/pkg/auth/manager/publicshares"
 	// Add your own here
